+++ app/src/main/java/br/senai/sp/fitware/model/UserToken.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class UserToken(
     @SerializedName("id")
-    var userId: Int = 0,
+    var userId: Long = 0,
     var userPerfil: String = "",
     var userName: String = ""
 )
